Update cache to use inverse rate if available
CachingRateProvider will now return a rate calculated from the inverse rate whenever possible.

The inverse rate is never persisted, however. This approach results in a potentially redundant SQL call and it would be more efficient to store both the actual and the inverse value under the same normalized key, but I followed the specification given and opted for two SQL calls instead.